requirements.txt
Django>=4.2,<5.0
django-cookie-consent==0.6.0
django-csp==3.8
django-ddm==2.0b7
djangorestframework==3.15.2
gunicorn==23.0.0
pandas==2.2.3
psycopg[binary]==3.2.4
python-dotenv==1.0.1
requests==2.32.3
